Remove membership from the admins "New Member Email" (#31)
Currently the admin version of the New Member Email goes to both the
membership@kwartzlab.ca email and the bod@kwartzlab.ca email. As we move
towards sharing access to the membership email with a select few
members, sensitive data included in the admin version of the email
shouldn't be included.