feat: hide trainer names in the tool directory if they are not active members (#37)
Fixes issue #3 by removing inactive trainers from the list of available
tool trainers. This can also be implemented by showing the trainer name
but with a different label colour. I chose not to do it that way because
a) it adds additional complexity and b) it provides a level of
visibility into member statuses that could and maybe should very well be
private.