chore(deploy): Current deploy script (#44)
This is a first stab at a deploy script. It follows the Laravel docs
and attempts to be reasonably fault tolerant.
This is not result in a no downtime deploy, which in kOS's current
state is acceptable.

This is what I just used to deploy. I'm sure it can be improved but it
seemed to have worked. I ran it using `sudo -u www-data ./deploy.sh`. In
order to do this I created a deploy ssh key and added it to the repo

refs:
- https://laravel.com/docs/10.x/deployment
- https://hakk.dev/blog/posts/www-data-git-deployment/