chore(sec): fix database connection after enabling ssl (#94)
I enforced SSL on the mysql server since the BOD asked for access.
However, this caused our application connection to lose connection.
Adding a UNIX socket connection mirrors the way the `mysql-client`
accesses the database.

This change is currently on the server as a hotfix. DO NOT DEPLOY until
this is merged!