feat: Add Revoke All Access to Tool button for BoD members (#93)
Adds a button and modal in the gatekeeper dashboards which revokes all
user access to that gatekeeper.
Currently the button is located just under the table of authorized
users, and only accessible to BoD members.
Resolves #88.

I'm keeping this as a draft for now because I wasn't able to properly
run the lint/pint test, for some reason it 'fixed' issues in scripts
that were not part of this feature. And also because I want to confirm
the location of the button before this is merged. Not to mention that
this is a fairly destructive process, since currently there's no way to
retrieve authorization after it's revoked. However the feature is fully
functional.


![image](https://github.com/user-attachments/assets/4589e51f-b9b5-432a-8433-4ddde3f869d8)

---------

Co-authored-by: Verdi R-D <verdi@azend.org>