Added 'autocomplete' to kiosk password input to stop autofill (#118)
Currently when on the kiosk page, it attempts to autofill the password
saved for kos.kwartzlab.ca. This is causing issues with the fob reader,
since the user will have to manually delete the filled password without
the box being visible.

Since there's no reason for us to have autofill on this page, setting
`autocomplete="new-password"` should stop the from happening